Changes how Google identifies the user.
The Google OAuth2 provider uses the account ID instead of the Google
plus URL. This allows people without Google plus account to log in.

This is a breaking change. To migrate, run an SQL query that removes the
`https://plus.google.com/` prefix in the authid field of the auth table.

This is a sample SQL query:
	UPDATE auth SET
		authid = REPLACE(authid, 'https://plus.google.com/', '')
	WHERE provider = 'google';